test fixes for integer mode

+++ src/exercises/add_1.c
@@ -1,8 +1,8 @@
 int add(int x, int y)
 /* --BEGIN-- */
-/*@ requires let Sum = (i64) x + (i64) y;
-             -2147483648i64 <= Sum; Sum <= 2147483647i64;
-    ensures return == (i32) Sum;
+/*@ requires let Sum = x + y;
+             -2147483648 <= Sum; Sum <= 2147483647;
+    ensures return == Sum;
 @*/
 /* --END-- */
 {
